Przygotowanie do searcha
