+++ Spizarnia/src/app/components/pages/grocery-list/grocery-list-view/grocery-list-view.component.html
@@ -1,75 +1,84 @@
-<div class="grocery-list-view-container">
-  <table mat-table [dataSource]="dataSource" matSort>
-    <!-- Columns for id, name, quantity, category, price, and action buttons -->
-    
-    <ng-container matColumnDef="id">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>ID</th>
-      <td mat-cell *matCellDef="let product">{{ product.id }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Nazwa</th>
-      <td mat-cell *matCellDef="let product">{{ product.products.name }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="quantityOfProduct">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Ilość produktu</th>
-      <td mat-cell *matCellDef="let product">{{ product.products.quantity }} {{ product.products.unit }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="quantity">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Ilość</th>
-      <td mat-cell *matCellDef="let product">{{ product.quantity }} {{ product.unit }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="categoryName">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Kategoria</th>
-      <td mat-cell *matCellDef="let product">{{ product.products.category.categoryName }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="price">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Cena</th>
-      <td mat-cell *matCellDef="let product">{{ product.products.price }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="totalPrice">
-      <th mat-header-cell *matHeaderCellDef mat-sort-header>Suma</th>
-      <td mat-cell *matCellDef="let product">{{ product.products.price*product.quantity }}</td>
-    </ng-container>
-
-    <ng-container matColumnDef="edit">
-      <th mat-header-cell *matHeaderCellDef></th>
-      <td mat-cell *matCellDef="let product">
-        <button class="action-edit-button" (click)="editProduct(product.id)">Edytuj</button>
-      </td>
-    </ng-container>
-
-    <ng-container matColumnDef="delete">
-      <th mat-header-cell *matHeaderCellDef></th>
-      <td mat-cell *matCellDef="let product">
-        <button class="action-edit-button" (click)="deleteProductModelFromCart(product.id)">Usuń</button>
-      </td>
-    </ng-container>
-
-    <!-- Rows -->
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
-    <!-- Total Price Row -->
-
-    <!-- <tr>
-      <td mat-cell colspan="7"></td> 
-      <td mat-header-cell><strong>Suma:</strong></td>
-      <td mat-cell><strong>{{ totalSummaryPrice | currency: 'PLN' }}</strong></td>
-      <td mat-cell></td> 
-    </tr> -->
-  </table>
-
-  {{ totalSummaryPrice}}
-
-  @if(dataSource.data.length === 0){
-  <div class="no-data">
-    Brak produktów w spiżarni.
+<div class="header-container">
+  <div class="title">Lista Zakupów</div>
+  <div class="search-bar">
+    <input type="text"
+           placeholder="Wyszukaj produkt..."
+           [(ngModel)]="searchTerm"
+           (input)="onSearch()" />
+    <button class="action-button" (click)="onSearch()">Szukaj</button>
+    <button class="action-button">Usuń listę zakupów</button>
   </div>
-  }  
+</div>
 
+<!-- Wyświetlanie dynamicznego widoku -->
+<div class="site-content">
+  <!-- Table view to show the grocery list -->
+  <div class="grocery-list-view-container">
+    <table mat-table [dataSource]="dataSource" matSort>
+      <!-- Columns for id, name, quantity, category, price, and action buttons -->
+
+      <ng-container matColumnDef="id">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>ID</th>
+        <td mat-cell *matCellDef="let product">{{ product.id }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="name">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Nazwa</th>
+        <td mat-cell *matCellDef="let product">{{ product.products.name }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="quantityOfProduct">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Ilość produktu</th>
+        <td mat-cell *matCellDef="let product">{{ product.products.quantity }} {{ product.products.unit }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="quantity">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Ilość</th>
+        <td mat-cell *matCellDef="let product">{{ product.quantity }} {{ product.unit }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="categoryName">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Kategoria</th>
+        <td mat-cell *matCellDef="let product">{{ product.products.category.categoryName }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="price">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Cena</th>
+        <td mat-cell *matCellDef="let product">{{ product.products.price }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="totalPrice">
+        <th mat-header-cell *matHeaderCellDef mat-sort-header>Suma</th>
+        <td mat-cell *matCellDef="let product">{{ product.products.price * product.quantity }}</td>
+      </ng-container>
+
+      <ng-container matColumnDef="edit">
+        <th mat-header-cell *matHeaderCellDef></th>
+        <td mat-cell *matCellDef="let product">
+          <button class="action-edit-button" (click)="editProduct(product.id)">Edytuj</button>
+        </td>
+      </ng-container>
+
+      <ng-container matColumnDef="delete">
+        <th mat-header-cell *matHeaderCellDef></th>
+        <td mat-cell *matCellDef="let product">
+          <button class="action-edit-button" (click)="deleteProductModelFromCart(product.id)">Usuń</button>
+        </td>
+      </ng-container>
+
+      <!-- Rows -->
+      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+      <tr mat-row *matRowDef="let row; columns: displayedColumns;"></tr>
+    </table>
+
+    <!-- Display total summary price -->
+    <div class="total-summary">
+      <strong>Suma:</strong> {{ totalSummaryPrice | currency: 'PLN' }}
+    </div>
+
+    <!-- If no data is available -->
+    <div *ngIf="dataSource.data.length === 0" class="no-data">
+      Brak produktów w spiżarni.
+    </div>
+  </div>
 </div>
